+++ src/main/java/gift/entity/Gift.java
@@ -1,3 +1,3 @@
-package gift;
+package gift.entity;
 
 public record Gift(long id, String name, int price, String imageUrl) {}
